Add single shot mode
This is useful for environments such as OpenShift/K8s where it is natural
to use CronJob resources to run periodic tasks such as this.

It is more efficient on cluster resources to run it this way without
sacrificing any timeliness of data collection.

The single shot mode is activated by setting the collection_interval
to 0.