Add Graceful Shutdown
Ensures that a final collection cycle is run upon receiving SIGTERM.
This is very useful in environments with ephemeral storage like
K8s/OpenShift where the lightspeed pods write to `emptyDir` volumes
for simplicity.

Data can still be lost in such cases if the final collection cycle
fails, but at least it tries.